% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentAnalysis.r
\name{sentAnalysis}
\alias{sentAnalysis}
\title{Returns the valence of the input Text}
\usage{
sentAnalysis(inputText = "")
}
\value{
valence value between -1 and 1
}
\description{
Returns the valence of the input Text
}
